Add a trait to parse the Curl response.
