+++ example_parallel.php
@@ -6,8 +6,8 @@ require "vendor/autoload.php";
 
 $httpClientParallel = new HttpClientParallel(
     new \ByJG\Util\HttpClient(),
-    function ($body, $id) {
-        echo "[$id] => $body\n";
+    function ($response, $id) {
+        echo "[$id] => " . $response->getStatusCode() . " - " . $response->getBody() . "\n";
     },
     function ($error, $id) {
         echo "[$id] => $error\n";
@@ -19,7 +19,7 @@ $uri1 = \ByJG\Util\Uri::getInstanceFromString('http://www.byjg.com.br/ws/cep?htt
 $request1 = \ByJG\Util\Psr7\Request::getInstance($uri1)->withMethod('POST');
 
 $uri2 = \ByJG\Util\Uri::getInstanceFromString('http://www.byjg.com.br/ws/cep?httpmethod=obterLogradouro&cep=30130000');
-$request2 = \ByJG\Util\Psr7\Request::getInstance($uri2)->withMethod('POST');
+$request2 = \ByJG\Util\Psr7\Request::getInstance($uri2)->withMethod('GET');
 
 $httpClientParallel
     ->addRequest($request1)
